requirements.txt
alabaster==0.7.12
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
Babel==2.10.1
backports.zoneinfo==0.2.1
bcrypt==3.2.2
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dateparser==1.1.1
docutils==0.17.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.8
PyPDF2==1.27.12
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
reportlab==3.6.9
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.36
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uuid==1.30
Werkzeug==2.1.2
zipp==3.8.0
